requirements.txt
streamlit
yfinance
pandas
numpy
openpyxl
psycopg2-binary
ta
